fix(tun): Fix memory leak
TCP connections sometimes do not closed by either 2 FIN+ACK or RST. Such
as when the client device crashes or switches to a different network.

In such case, the proxy server still sends about 8 FIN or PSH packets
with the same Seq. This commit detects that and clean the connection
state in memory.

It also removes the mock IP allocation logic. Because all clients should
always share one IP, that is the IP of the TUN device. Kernal already
does the NAT for translating the real client IP to the TUN IP.